requirements.txt
streamlit==1.28.1
sentence-transformers==2.2.2
faiss-cpu==1.13.1
pandas==2.2.0
numpy==1.26.4
pyyaml==6.0.1
scikit-learn==1.4.0
chromadb==0.4.22
plotly==5.18.0
transformers==4.37.2
torch==2.1.2
tqdm==4.66.1